eagerly consume the hyper request
To pave the way for support of `conduit::Request#body()` we need to
eagerly consume the hyper request and extract the data we need. Taking
the body of a `hyper::Request` now consumes said request and yields a
stream we can block on.

Once we get access to this stream we can no longer use the original
request, so in order to implement the conduit API we need to save the
data before then.