schedule conduit app on a cpu pool
since the conduit application API is a blocking interface (i.e: it
doesn't reutrn a future), our event loop knows nothing about it. this
treats the handler like an opaque unit of work to be scheduled on a
thread pool.

this should free up the event loop to handle async I/O for the HTTP
server while the application is busy generating a response.